api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
azure-common==1.1.28
azure-core==1.37.0
azure-search-documents==11.6.0
azure-storage-blob==12.27.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
fastapi==0.128.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
jiter==0.12.0
lxml==6.0.2
openai==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.5.0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.21
requests==2.32.5
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
